Move Mui theme
