+++ src/app/styles.js
@@ -1,0 +1,29 @@
+import { createMuiTheme } from '@material-ui/core/styles';
+import { grey } from '@material-ui/core/colors';
+
+export const muiTheme = {
+  dark: createMuiTheme({
+    palette: {
+      type: 'dark',
+      primary: grey,
+      secondary: grey,
+    },
+    props: {
+      MuiButtonBase: {
+        disableRipple: true,
+      },
+    },
+  }),
+  light: createMuiTheme({
+    palette: {
+      type: 'light',
+      primary: grey,
+      secondary: grey,
+    },
+    props: {
+      MuiButtonBase: {
+        disableRipple: true,
+      },
+    },
+  }),
+};
