+++ src/components/layer-controller/styles.js
@@ -1,32 +1,4 @@
-import { makeStyles, createMuiTheme } from '@material-ui/core/styles';
-import { grey } from '@material-ui/core/colors';
-
-export const controllerTheme = {
-  dark: createMuiTheme({
-    palette: {
-      type: 'dark',
-      primary: grey,
-      secondary: grey,
-    },
-    props: {
-      MuiButtonBase: {
-        disableRipple: true,
-      },
-    },
-  }),
-  light: createMuiTheme({
-    palette: {
-      type: 'light',
-      primary: grey,
-      secondary: grey,
-    },
-    props: {
-      MuiButtonBase: {
-        disableRipple: true,
-      },
-    },
-  }),
-};
+import { makeStyles } from '@material-ui/core/styles';
 
 export const useOptionStyles = makeStyles(theme => ({
   paper: {
